move all methods except playGame to a new class GameIO
